templates/python-with-api/pyproject.toml
[project]
name = "python-with-api"
version = "0.1.0"
description = "A beluga labs template for a Python project with an API"
authors = [
    { name = "beluga labs" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "SQLAlchemy==2.0.40",
    "python-dotenv==1.1.0",
    "pydantic==2.11.3",
    "alembic==1.15.2",
    "pytest==8.3.5",
    "httpx",
    "slowapi",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
]

[project.urls]
homepage = "https://belugalabs.digital"
repository = "https://github.com/beluga-labs/beluga-templates"

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["venv", "env", "alembic"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
strict = true
exclude = "venv|env|alembic"

[tool.pdm]
distribution = false
